Agent query: J'ai corrigé les méthodes dupliquées dans ReportBuilder et les icônes dans MainViewSwing. Il reste encore quelques ajustements à faire. Souhaitez-vous que je continue avec les corrections ou préférez-vous que nous prenions une approche différente pour la mise à jour du module de rapports ?
Update dependencies, improve report generation with Apache POI charts, and fix data type inconsistencies.

+++ src/main/java/com/poissonnerie/controller/VenteController.java
@@ -364,7 +364,7 @@ public class VenteController {
 
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
             pstmt.setInt(1, venteId);
-            pstmt.setInt(2, ligne.getProduit().getId());
+            pstmt.setLong(2, ligne.getProduit().getId()); // Modifié de setInt à setLong pour correspondre au type Long
             pstmt.setInt(3, ligne.getQuantite());
             pstmt.setDouble(4, ligne.getPrixUnitaire());
 
@@ -378,7 +378,7 @@ public class VenteController {
 
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
             pstmt.setInt(1, ligne.getQuantite());
-            pstmt.setLong(2, ligne.getProduit().getId());
+            pstmt.setLong(2, ligne.getProduit().getId()); // Modifié de setInt à setLong
             pstmt.setInt(3, ligne.getQuantite());
 
             int rowsAffected = pstmt.executeUpdate();
